requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.3.0
jira==3.5.1
slack_sdk==3.21.3
elasticsearch==8.9.0
thehive4py==1.8.1
